Differential abundance

This function returns the fold change and associated p value for a differential abundance analysis comparing samples in two groups.

Usage

internalDiffAbund(
  data,
  comparator,
  method = c("DESeq2", "Maaslin2"),
  pValueFloor = P_VALUE_FLOOR,
  verbose = c(TRUE, FALSE)
)

# S4 method for class 'AbundanceData,Comparator'
internalDiffAbund(
  data,
  comparator,
  method = c("DESeq2", "Maaslin2"),
  pValueFloor = P_VALUE_FLOOR,
  verbose = c(TRUE, FALSE)
)

Arguments

data

AbsoluteAbundanceData object

comparator

Comparator object specifying the variable and values or bins to be used in dividing samples into groups.

method

string defining the the differential abundance method. Accepted values are 'DESeq2' and 'Maaslin2'.

pValueFloor

numeric value that indicates the smallest p value that should be returned. The corresponding adjusted p value floor will also be updated based on this value, and will be set to the maximum adjusted p value of all floored p values. The default value uses the P_VALUE_FLOOR=1e-200 constant defined in this package.

verbose

boolean indicating if timed logging is desired

Value

ComputeResult object